+++ src/symbols.js
@@ -2,6 +2,7 @@ export default {
 	open: '[',
 	close: ']',
 	path: '.',
+	opt: '?',
 	append: '|',
 	param: ':'
 };