+++ CHANGES.md
@@ -72,3 +72,4 @@ It is still possible to run matchdom without a browser environment, as long as n
 - add `split`, `join`, `slice` filters
 - replace ternary '?' by 'alt'
 - allow one or two parameters for 'alt'
+- paths can use optional chaining to force undefined to be evaluated as null